pyproject.toml
[project]
name = "spatter-monitoring"
version = "0.1.0"
description = "Production-grade MLOps pipeline for 3D printing spatter classification"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Huang De Jun", email = "dejun.huang.sg@gmail.com"}
]
dependencies = [
    "h5py>=3.15.1",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
# Training dependencies
training = [
    "torch",
    "torchvision",
    "pillow",
    "tqdm",
]

# MLOps dependencies (Phase 1)
mlops = [
    "mlflow",
    "hydra-core",
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
]

# HPO dependencies (Phase 2)
hpo = [
    "optuna",
    "ray[tune]",
    "tensorboard",
]

# API dependencies (Phase 1)
api = [
    "fastapi",
    "uvicorn",
    "pydantic",
]

# Development dependencies
dev = [
    "jupyter",
    "ipykernel",
    "black",
    "isort",
]

# All dependencies
all = [
    "spatter-monitoring[training,mlops,hpo,api,dev]",
]

[project.scripts]
spatter-extract = "spatter.extraction.pipeline:main"
spatter-train = "spatter.training.trainer:main"
spatter-predict = "spatter.inference.predictor:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spatter"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/spatter --cov-report=html --cov-report=term"

[tool.ruff]
line-length = 100
target-version = "py313"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
